Refactor top-level tool code
